+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Method `fallback()`
+- Equality operator support
+
+### Changed
+- Method  now accepts a value instead of a function. The old method is available as 
+- Method `chain()` now accepts a value instead of a function. The old method is available as `fallback()`
+
+### Removed
+- Method `filter()`
+
 ## [0.1.9] - 2020-06-27
 ### Changed
 - ifPresent() and ifNothing() return Maybe<T> instead of null